+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] / 2025-07-18
+### Features
+- Support `AddSlideOut` extension.
+- Support `SetDialogLauncher` and `GetDialogLauncher` extensions.
+### UI
+- Update `RibbonExtension` to support `AddSlideOut`.
+- Update `RibbonExtension` to support `SetDialogLauncher` and `GetDialogLauncher`.
+
 ## [0.3.0] / 2025-07-15
 ### Features
 - Support `CreateSplitButton` extension.
@@ -46,6 +54,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `Busy`, `Runtime`, `Tasks`, and `Windows` utilities.
 
 [vNext]: ../../compare/1.0.0...HEAD
+[0.4.0]: ../../compare/0.3.0...0.4.0
 [0.3.0]: ../../compare/0.2.0...0.3.0
 [0.2.0]: ../../compare/0.1.0...0.2.0
 [0.1.0]: ../../compare/0.1.0